added service layer tests

+++ models/burritoModel.js
@@ -14,7 +14,7 @@ module.exports = function (sequelize, DataTypes) {
       allowNull: false
     },
     price: {
-      type: DataTypes.FLOAT(11, 2),
+      type: DataTypes.FLOAT,
       allowNull: false
     }
   });
